fix: resolve TDZ crashes and circular dependency in ballerina-core   - Reorder libraryProjectPaths useMemo before its first use in WorkspaceOverview   - Move module-level consts inside findScopeByModule to prevent webpack scope-hoisting TDZ   - Break state-machine-types ↔ extended-lang-client circular dep via new shared-types.ts   - Deduplicate findScopeByModule by importing from ballerina-core in the visualizer
