+++ workspaces/ballerina/ballerina-core/src/interfaces/extended-lang-client.ts
@@ -29,7 +29,7 @@ import { SqFlow } from "../rpc-types/sequence-diagram/interfaces";
 import { FieldType, FunctionModel, ListenerModel, ServiceClassModel, ServiceInitModel, ServiceModel } from "./service";
 import { CDModel } from "./component-diagram";
 import { DMModel, ExpandedDMModel, IntermediateClause, Mapping, VisualizableField, FnMetadata, ResultClauseType, IOType } from "./data-mapper";
-import { ArtifactData, DataMapperMetadata, SCOPE } from "../state-machine-types";
+import { ArtifactData, DataMapperMetadata, SCOPE } from "./shared-types";
 import { ToolParameters } from "../rpc-types/ai-agent/interfaces";
 
 export interface DidOpenParams {
