+++ workspaces/ballerina/ballerina-core/src/interfaces/shared-types.ts
@@ -1,0 +1,39 @@
+/**
+ * Copyright (c) 2026, WSO2 LLC. (https://www.wso2.com) All Rights Reserved.
+ *
+ * WSO2 LLC. licenses this file to you under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+import { CodeData, DIRECTORY_MAP } from "./bi";
+
+export enum SCOPE {
+    AUTOMATION = "automation",
+    INTEGRATION_AS_API = "integration-as-api",
+    EVENT_INTEGRATION = "event-integration",
+    FILE_INTEGRATION = "file-integration",
+    AI_AGENT = "ai-agent",
+    MCP = "mcp-server",
+    LIBRARY = "library",
+    ANY = "any"
+}
+
+export interface ArtifactData {
+    artifactType: DIRECTORY_MAP;
+    identifier?: string;
+}
+
+export interface DataMapperMetadata {
+    name: string;
+    codeData: CodeData;
+}
