+++ workspaces/ballerina/ballerina-core/src/utils/identifier-utils.ts
@@ -21,16 +21,13 @@ import { ComponentInfo } from "../interfaces/ballerina";
 import { BallerinaProjectComponents } from "../interfaces/extended-lang-client";
 import { SCOPE } from "../state-machine-types";
 
-const INTEGRATION_API_MODULES = ["http", "graphql", "tcp"];
-const EVENT_INTEGRATION_MODULES = ["kafka", "rabbitmq", "salesforce", "trigger.github", "mqtt", "asb"];
-const FILE_INTEGRATION_MODULES = ["ftp", "file"];
-const AI_AGENT_MODULE = "ai";
-const MCP_MODULE = "mcp";
-
 export function findScopeByModule(moduleName: string): SCOPE {
-    if (AI_AGENT_MODULE === moduleName) {
+    const INTEGRATION_API_MODULES = ["http", "graphql", "tcp"];
+    const EVENT_INTEGRATION_MODULES = ["kafka", "rabbitmq", "salesforce", "trigger.github", "mqtt", "asb", "mssql", "mysql", "postgresql"];
+    const FILE_INTEGRATION_MODULES = ["ftp", "file"];
+    if (moduleName === "ai") {
         return SCOPE.AI_AGENT;
-    } else if (MCP_MODULE === moduleName) {
+    } else if (moduleName === "mcp") {
         return SCOPE.MCP;
     } else if (INTEGRATION_API_MODULES.includes(moduleName)) {
         return SCOPE.INTEGRATION_AS_API;
